+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Bug Fixes**:
 
 - Add `sentry.dsc.transaction` and `sentry.dsc.trace_id` to all v2 spans. ([#6001](https://github.com/getsentry/relay/pull/6001))
+- Obtain PII values for `SpanData` fields from `sentry-conventions`. ([#5997](https://github.com/getsentry/relay/pull/5997))
 
 ## 26.5.0